+++ CHANGELOG.md
@@ -4,8 +4,9 @@
 
 ### Changes
 
+- Integrated miden-client (#11).
 - Redesigned the home frontend (#20).
 
 ## 0.10.0 (2025-07-10)
 
-For previous changes see https://github.com/0xMiden/miden-node/blob/main/CHANGELOG.md
+For previous changes see the changelog in https://github.com/0xMiden/miden-node.